refactor(BitcoinExplorers): moved up API files
