+++ src/explorers/index.js
@@ -1,12 +1,12 @@
-import { getEtherScanFetcher } from './ethereum';
-import { getBitcoinTransactionFromApi } from './bitcoin/bitcoin-explorer';
-import { TRANSACTION_APIS } from '../constants';
+import { getEtherScanFetcher } from '../ethereum';
+import { getBlockchainTransactionFromApi } from '../blockchain-explorer';
+import { TRANSACTION_APIS } from '../../constants';
 
 const BitcoinExplorers = [
-  (transactionId, chain) => getBitcoinTransactionFromApi(TRANSACTION_APIS.Blockcypher, transactionId, chain),
-  (transactionId, chain) => getBitcoinTransactionFromApi(TRANSACTION_APIS.Bitpay, transactionId, chain),
-  (transactionId, chain) => getBitcoinTransactionFromApi(TRANSACTION_APIS.Blockexplorer, transactionId, chain),
-  (transactionId, chain) => getBitcoinTransactionFromApi(TRANSACTION_APIS.Blockstream, transactionId, chain)
+  (transactionId, chain) => getBlockchainTransactionFromApi(TRANSACTION_APIS.Blockcypher, transactionId, chain),
+  (transactionId, chain) => getBlockchainTransactionFromApi(TRANSACTION_APIS.Bitpay, transactionId, chain),
+  (transactionId, chain) => getBlockchainTransactionFromApi(TRANSACTION_APIS.Blockexplorer, transactionId, chain),
+  (transactionId, chain) => getBlockchainTransactionFromApi(TRANSACTION_APIS.Blockstream, transactionId, chain)
 ];
 
 const EthereumExplorers = [
@@ -15,7 +15,7 @@ const EthereumExplorers = [
 
 // for legacy (pre-v2) Blockcerts
 const BlockchainExplorersWithSpentOutputInfo = [
-  (transactionId, chain) => getBitcoinTransactionFromApi(TRANSACTION_APIS.Blockcypher, transactionId, chain)
+  (transactionId, chain) => getBlockchainTransactionFromApi(TRANSACTION_APIS.Blockcypher, transactionId, chain)
 ];
 
 export { BitcoinExplorers, EthereumExplorers, BlockchainExplorersWithSpentOutputInfo };
